requirements.txt
Click==7.0
gunicorn==19.7.1
Flask==1.0.2
itsdangerous==1.1.0
Jinja2==2.10
Pillow==5.3.0
MarkupSafe==1.1.0
pytesseract==0.2.5
Werkzeug==0.14.1
